requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
blis==1.1.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.10
diskcache==5.6.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
llama_cpp_python==0.2.20
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
murmurhash==1.0.11
namex==0.0.8
numpy==1.26.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
preshed==3.0.9
protobuf==5.29.2
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
PyYAML==6.0.2
rdflib==7.1.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SPARQLWrapper==2.0.0
srsly==2.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
thinc==8.3.3
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
